move vtext tests
