+++ yew/src/virtual_dom/smr.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "sans_mount_render")]
+
 use htmlescape;
 use std::convert::TryFrom;
 use thiserror::Error as ThisError;
@@ -162,5 +164,62 @@ impl TryFrom<VNode> for Html {
             VNode::VList(vlist) => Html::try_from(vlist)?,
             VNode::VRef(_) => Err(HtmlRenderError::UnserializableVRef)?,
         })
+    }
+}
+
+
+#[cfg(test)]
+mod test_vtext {
+    use crate::html;
+    use super::Html;
+    use std::convert::TryFrom;
+
+    #[cfg(feature = "wasm_test")]
+    use wasm_bindgen_test::{wasm_bindgen_test as test, wasm_bindgen_test_configure};
+
+    #[cfg(feature = "wasm_test")]
+    wasm_bindgen_test_configure!(run_in_browser);
+
+    #[test]
+    #[cfg(feature = "sans_mount_render")]
+    fn text_as_root_smr() {
+        let a = html! {
+            "Text Node As Root"
+        };
+
+        let b = html! {
+            { "Text Node As Root" }
+        };
+
+        assert_eq!(
+            Html::try_from(a.clone()).expect("HTML stringify error"),
+            Html::try_from(b.clone()).expect("HTML stringify error")
+        );
+        assert!(
+            Html::try_from(b).expect("HTML stringify error").to_string() == "Text Node As Root"
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "sans_mount_render")]
+    fn special_chars_smr() {
+        let a = html! {
+            "some special-chars\"> here!"
+        };
+
+        let b = html! {
+            { "some special-chars\"> here!" }
+        };
+
+        assert_eq!(
+            Html::try_from(a.clone()).expect("HTML stringify error"),
+            Html::try_from(b.clone()).expect("HTML stringify error")
+        );
+        assert_eq!(
+            Html::try_from(b.clone())
+                .expect("HTML stringify error")
+                .to_string(),
+            "some special-chars&quot;&gt; here!"
+        );
     }
 }
